+++ packages/react-components/src/card/components/short-story.js
@@ -3,42 +3,35 @@ import PropTypes from 'prop-types'
 import styled from 'styled-components'
 // component
 import Image from '../../image-with-fallback'
-import { P1, P2, P3 } from '../../text/paragraph'
-import { H4 } from '../../text/headline'
-import { Bookmark } from '../../icon'
-import { TextButton } from '../../button'
+import { P3 } from '../../text/paragraph'
+import { H6 } from '../../text/headline'
 import Link from '../../customized-link'
 // enum
 import { Size } from '../../shared-enum'
 // @twreporter
 import { colorGrayscale } from '@twreporter/core/lib/constants/color'
+import { ARTICLE_THEME } from '@twreporter/core/lib/constants/theme'
+import entityPaths from '@twreporter/core/lib/constants/entity-path'
 import {
   BRANCH,
   BRANCH_PROP_TYPES,
 } from '@twreporter/core/lib/constants/release-branch'
-import { ARTICLE_THEME } from '@twreporter/core/lib/constants/theme'
-import entityPaths from '@twreporter/core/lib/constants/entity-path'
 
 const imageStyle = {
   width: {
-    [Size.S]: '72px',
-    [Size.L]: '216px',
+    [Size.S]: '48px',
+    [Size.L]: '80px',
   },
   height: {
-    [Size.S]: '72px',
-    [Size.L]: '144px',
+    [Size.S]: '48px',
+    [Size.L]: '80px',
   },
   marginLeft: {
     [Size.S]: '8px',
-    [Size.L]: '32px',
+    [Size.L]: '16px',
   },
 }
-const bookmarkStyle = {
-  marginTop: {
-    [Size.S]: '16px',
-    [Size.L]: '8px',
-  },
-}
+
 const metaStyle = {
   marginBottom: {
     [Size.S]: '4px',
@@ -49,9 +42,7 @@ const metaStyle = {
 const Container = styled(Link)`
   text-decoration: none;
   &:hover {
-    .hover {
-      opacity: 0.8;
-    }
+    opacity: 0.8;
   }
 `
 
@@ -78,22 +69,7 @@ const Meta = styled(FlexGroup)`
   }
   ${props => (props.hide ? `display: none;` : '')}
 `
-const DescContainer = styled.div`
-  color: ${colorGrayscale.gray800};
-  margin-top: 8px;
-  p {
-    display: -webkit-box;
-    text-overflow: ellipsis;
-    overflow: hidden;
-    line-clamp: 3;
-    -webkit-line-clamp: 3;
-    -webkit-box-orient: vertical;
-  }
-`
-const BookmarkContainer = styled(FlexGroup)`
-  align-self: flex-end;
-  margin-top: ${props => bookmarkStyle.marginTop[props.size]};
-`
+
 const ImageContainer = styled(FlexGroup)`
   flex: 0 0 auto;
   align-self: center;
@@ -104,38 +80,24 @@ const ImageContainer = styled(FlexGroup)`
 const LeftColumn = styled(FlexGroupColumn)`
   flex: 1;
 `
-const TitleText = styled(H4)`
+const TitleText = styled(H6)`
   color: ${colorGrayscale.gray800};
 `
 
-const ArticleCard = ({
+const ShortStory = ({
   title = '',
-  description = '',
   date = '',
   image = {},
   category = '',
   size = Size.S,
-  isBookmarked = false,
-  toggleBookmark,
   releaseBranch = BRANCH.master,
-  showIsBookmarked = false,
   style = ARTICLE_THEME.v2.default,
   slug = '',
 }) => {
   const hideMeta = !category && !date
-  const titleJSX = title ? (
-    <TitleText text={title} type="article" className="hover" />
-  ) : null
-  const dateJSX = date ? <P3 text={date} className="hover" /> : null
-  const categoryJSX = category ? <P3 text={category} className="hover" /> : null
-  const descriptionJSX = description ? (
-    size === Size.S ? (
-      <P2 text={description} className="hover" />
-    ) : (
-      <P1 text={description} className="hover" />
-    )
-  ) : null
-
+  const titleJSX = title ? <TitleText text={title} type="article" /> : null
+  const dateJSX = date ? <P3 text={date} /> : null
+  const categoryJSX = category ? <P3 text={category} /> : null
   const isInteractiveArticle = style === ARTICLE_THEME.v2.interactive
   const link = {
     to: `${
@@ -146,39 +108,14 @@ const ArticleCard = ({
     target: isInteractiveArticle ? '_blank' : '',
   }
 
-  const onBookmarkClick = event => {
-    event.preventDefault()
-    toggleBookmark()
-  }
-
-  const bookmarkIcon = (
-    <Bookmark
-      type={isBookmarked ? Bookmark.Type.SAVED : Bookmark.Type.ADD}
-      releaseBranch={releaseBranch}
-    />
-  )
-  const bookmarkButton = (
-    <TextButton
-      theme={TextButton.THEME.normal}
-      style={TextButton.Style.LIGHT}
-      leftIconComponent={bookmarkIcon}
-      text={isBookmarked ? '已收藏' : '收藏'}
-    />
-  )
-  const bookmarkJSX = showIsBookmarked ? (
-    <BookmarkContainer onClick={onBookmarkClick} size={size}>
-      {bookmarkButton}
-    </BookmarkContainer>
-  ) : null
-
   const metaJSX = (
-    <Meta hide={hideMeta} size={size} className="hover">
+    <Meta hide={hideMeta} size={size}>
       {categoryJSX}
       {dateJSX}
     </Meta>
   )
   const imageJSX = (
-    <ImageContainer size={size} className="hover">
+    <ImageContainer size={size}>
       <Image src={image.src} alt={image.alt} releaseBranch={releaseBranch} />
     </ImageContainer>
   )
@@ -194,10 +131,6 @@ const ArticleCard = ({
             </FlexGroupColumn>
             <FlexGroup>{imageJSX}</FlexGroup>
           </FlexSpaceBetween>
-          <FlexGroupColumn>
-            <DescContainer>{descriptionJSX}</DescContainer>
-            {bookmarkJSX}
-          </FlexGroupColumn>
         </FlexGroupColumn>
       </Container>
     )
@@ -209,20 +142,15 @@ const ArticleCard = ({
       <FlexSpaceBetween>
         <LeftColumn>
           {metaJSX}
-          <FlexGroupColumn>
-            {titleJSX}
-            <DescContainer>{descriptionJSX}</DescContainer>
-            {bookmarkJSX}
-          </FlexGroupColumn>
+          <FlexGroupColumn>{titleJSX}</FlexGroupColumn>
         </LeftColumn>
         {imageJSX}
       </FlexSpaceBetween>
     </Container>
   )
 }
-ArticleCard.propTypes = {
+ShortStory.propTypes = {
   title: PropTypes.string.isRequired,
-  description: PropTypes.string.isRequired,
   date: PropTypes.string,
   image: PropTypes.shape({
     alt: PropTypes.string,
@@ -231,13 +159,10 @@ ArticleCard.propTypes = {
   }),
   category: PropTypes.string,
   size: PropTypes.oneOf(Object.values(Size)),
-  isBookmarked: PropTypes.bool,
-  toggleBookmark: PropTypes.func,
   releaseBranch: BRANCH_PROP_TYPES,
-  showIsBookmarked: PropTypes.bool,
   style: PropTypes.oneOf(Object.values(ARTICLE_THEME.v2)),
   slug: PropTypes.string,
 }
-ArticleCard.Size = Size
+ShortStory.Size = Size
 
-export default ArticleCard
+export default ShortStory