+++ packages/react-components/src/hook/index.js
@@ -1,9 +1,13 @@
 import useOutsideClick from './use-outside-click'
 import useFontFaceObserver from './use-font-face-observer'
+import useBookmark from './use-bookmark'
+import useStore from './use-store'
 
-export { useOutsideClick, useFontFaceObserver }
+export { useOutsideClick, useFontFaceObserver, useBookmark, useStore }
 
 export default {
   useOutsideClick,
   useFontFaceObserver,
+  useBookmark,
+  useStore,
 }