+++ packages/react-components/src/listing-page/components/card-list.js
@@ -5,19 +5,16 @@ import styled from 'styled-components'
 import { ArticleCard } from '../../card'
 import FetchingWrapper from '../../is-fetching-wrapper'
 import Divider from '../../divider'
-import Link from '../../customized-link'
 import { DesktopAndAbove, TabletAndBelow } from '../../rwd'
 // constants
 import mockup from '../constants/mockup-spec'
 // @twreporter
-import entityPaths from '@twreporter/core/lib/constants/entity-path'
 import mq from '@twreporter/core/lib/utils/media-query'
 import { date2yyyymmdd } from '@twreporter/core/lib/utils/date'
 import {
   BRANCH,
   BRANCH_PROP_TYPES,
 } from '@twreporter/core/lib/constants/release-branch'
-import { ARTICLE_THEME } from '@twreporter/core/lib/constants/theme'
 // lodash
 import forEach from 'lodash/forEach'
 import get from 'lodash/get'
@@ -34,16 +31,9 @@ const Card = styled(ArticleCard)`
 const Container = styled.div`
   display: flex;
   flex-direction: column;
-  a {
-    color: inherit;
-    text-decoration: none;
-  }
 `
 const Item = styled.div`
   margin-bottom: 24px;
-  &:hover {
-    opacity: 0.7;
-  }
   &:last-child {
     margin-bottom: 0;
   }
@@ -53,7 +43,8 @@ const StyledDivider = styled(Divider)`
   margin-top: 24px;
 `
 const FlexItem = styled.div`
-  width: ${mockup.hd.maxWidth}px;
+  width: ${props =>
+    props.width !== 0 ? `${props.width}%` : mockup.hd.maxWidth}px;
   margin: 0 auto;
   display: flex;
   justify-content: flex-start;
@@ -61,11 +52,13 @@ const FlexItem = styled.div`
   flex-wrap: wrap;
 
   ${mq.desktopOnly`
-    width: ${mockup.desktop.maxWidth}px;
+    width: ${props =>
+      props.width !== 0 ? `${props.width}%` : mockup.desktop.maxWidth}px;
   `}
 
   ${mq.tabletOnly`
-    width: ${mockup.tablet.maxWidth}px;
+    width: ${props =>
+      props.width !== 0 ? `${props.width}%` : mockup.tablet.maxWidth}px;
   `}
 
   ${mq.mobileOnly`
@@ -80,6 +73,8 @@ const CardList = ({
   isFetching = false,
   showSpinner = false,
   releaseBranch = BRANCH.master,
+  showIsBookmarked = false,
+  width = 0,
 }) => {
   if (!data || data.length === 0) {
     return null
@@ -99,15 +94,6 @@ const CardList = ({
 
   const listJSX = _.map(data, item => {
     const { id, title, slug, style } = item
-    const isInteractiveArticle = style === ARTICLE_THEME.interactive
-    const link = {
-      to: `${
-        isInteractiveArticle
-          ? entityPaths.interactiveArticle
-          : entityPaths.article
-      }${slug}`,
-      target: isInteractiveArticle ? '_blank' : '',
-    }
     const articleCardProps = {
       title,
       description: _.get(item, 'og_description', ''),
@@ -120,26 +106,36 @@ const CardList = ({
       category: getFirstCategory(_.get(item, 'category_set', [])),
       date: date2yyyymmdd(_.get(item, 'published_date'), '/'),
       releaseBranch,
+      style,
+      slug,
+      isBookmarked: _.get(item, 'is_bookmarked', false),
+      toggleBookmark: _.get(item, 'toggle_bookmark'),
     }
 
     return (
       <Item key={id}>
-        <Link {...link}>
-          <DesktopAndAbove>
-            <Card {...articleCardProps} size={ArticleCard.Size.L} />
-          </DesktopAndAbove>
-          <TabletAndBelow>
-            <Card {...articleCardProps} size={ArticleCard.Size.S} />
-          </TabletAndBelow>
-          <StyledDivider />
-        </Link>
+        <DesktopAndAbove>
+          <Card
+            {...articleCardProps}
+            showIsBookmarked={showIsBookmarked}
+            size={ArticleCard.Size.L}
+          />
+        </DesktopAndAbove>
+        <TabletAndBelow>
+          <Card
+            {...articleCardProps}
+            showIsBookmarked={showIsBookmarked}
+            size={ArticleCard.Size.S}
+          />
+        </TabletAndBelow>
+        <StyledDivider />
       </Item>
     )
   })
 
   return (
     <Container>
-      <Content isFetching={isFetching} showSpinner={showSpinner}>
+      <Content isFetching={isFetching} showSpinner={showSpinner} width={width}>
         {listJSX}
       </Content>
     </Container>
@@ -157,11 +153,15 @@ CardList.propTypes = {
       published_date: PropTypes.string.isRequired,
       slug: PropTypes.string.isRequired,
       style: PropTypes.string,
+      is_bookmarked: PropTypes.bool,
+      toggle_bookmark: PropTypes.func,
     })
   ),
   isFetching: PropTypes.bool,
   showSpinner: PropTypes.bool,
   releaseBranch: BRANCH_PROP_TYPES,
+  showIsBookmarked: PropTypes.bool,
+  width: PropTypes.number,
 }
 
 export default CardList